+++ sentry-android-okhttp/src/test/java/io/sentry/android/okhttp/SentryOkHttpInterceptorTest.kt
@@ -7,6 +7,8 @@ import io.sentry.Breadcrumb
 import io.sentry.Hint
 import io.sentry.HttpStatusCodeRange
 import io.sentry.IHub
+import io.sentry.Scope
+import io.sentry.ScopeCallback
 import io.sentry.SentryOptions
 import io.sentry.SentryTraceHeader
 import io.sentry.SentryTracer
@@ -27,6 +29,7 @@ import okhttp3.mockwebserver.SocketPolicy
 import org.mockito.kotlin.any
 import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.check
+import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
@@ -47,6 +50,7 @@ class SentryOkHttpInterceptorTest {
         val server = MockWebServer()
         lateinit var sentryTracer: SentryTracer
         lateinit var options: SentryOptions
+        lateinit var scope: Scope
 
         @SuppressWarnings("LongParameterList")
         fun getSut(
@@ -76,7 +80,9 @@ class SentryOkHttpInterceptorTest {
                 }
                 isSendDefaultPii = sendDefaultPii
             }
+            scope = Scope(options)
             whenever(hub.options).thenReturn(options)
+            doAnswer { (it.arguments[0] as ScopeCallback).run(scope) }.whenever(hub).configureScope(any())
 
             sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
 
@@ -180,8 +186,18 @@ class SentryOkHttpInterceptorTest {
     }
 
     @Test
-    fun `when there is no active span, does not add sentry trace header to the request`() {
+    fun `when there is no active span, adds sentry trace header to the request from scope`() {
         val sut = fixture.getSut(isSpanActive = false)
+        sut.newCall(getRequest()).execute()
+        val recorderRequest = fixture.server.takeRequest()
+        assertNotNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])
+        assertNotNull(recorderRequest.headers[BaggageHeader.BAGGAGE_HEADER])
+    }
+
+    @Test
+    fun `when there is no active span and host if not allowed, does not add sentry trace header to the request`() {
+        val sut = fixture.getSut(isSpanActive = false)
+        fixture.options.setTracePropagationTargets(listOf("some-host-that-does-not-exist"))
         sut.newCall(getRequest()).execute()
         val recorderRequest = fixture.server.takeRequest()
         assertNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])