+++ sentry-spring-boot-starter-jakarta/src/main/java/io/sentry/spring/boot/jakarta/SentryAutoConfiguration.java
@@ -240,7 +240,6 @@ public class SentryAutoConfiguration {
       }
 
       @Bean
-      @Conditional(SentryTracingCondition.class)
       @ConditionalOnMissingBean(name = "sentryTracingFilter")
       public FilterRegistrationBean<SentryTracingFilter> sentryTracingFilter(
           final @NotNull IHub hub, final @NotNull TransactionNameProvider transactionNameProvider) {