+++ sentry/src/test/java/io/sentry/TransactionContextTest.kt
@@ -1,7 +1,7 @@
 package io.sentry
 
 import io.sentry.protocol.SentryId
-import io.sentry.protocol.TransactionNameSource
+import org.mockito.kotlin.mock
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -30,10 +30,14 @@ class TransactionContextTest {
     }
 
     @Test
-    fun `when context is created from trace header and baggage header, parent sampling decision of false is set from trace header`() {
-        val traceHeader = SentryTraceHeader(SentryId(), SpanId(), false)
-        val baggageHeader = Baggage.fromHeader("sentry-trace_id=a,sentry-transaction=sentryTransaction,sentry-sample_rate=0.3")
-        val context = TransactionContext.fromSentryTrace("name", TransactionNameSource.CUSTOM, "op", traceHeader, baggageHeader, null)
+    fun `when context is created from propagation context, parent sampling decision of false is set from trace header`() {
+        val logger = mock<ILogger>()
+        val propagationContext = PropagationContext.fromHeaders(
+            logger,
+            SentryTraceHeader(SentryId(), SpanId(), false).value,
+            "sentry-trace_id=a,sentry-transaction=sentryTransaction,sentry-sample_rate=0.3"
+        )
+        val context = TransactionContext.fromPropagationContext(propagationContext)
         assertNull(context.sampled)
         assertNull(context.profileSampled)
         assertFalse(context.parentSampled!!)
@@ -41,10 +45,14 @@ class TransactionContextTest {
     }
 
     @Test
-    fun `when context is created from trace header and baggage header, parent sampling decision of false is set from trace header if no sample rate is available`() {
-        val traceHeader = SentryTraceHeader(SentryId(), SpanId(), false)
-        val baggageHeader = Baggage.fromHeader("sentry-trace_id=a,sentry-transaction=sentryTransaction")
-        val context = TransactionContext.fromSentryTrace("name", TransactionNameSource.CUSTOM, "op", traceHeader, baggageHeader, null)
+    fun `when context is created from propagation context, parent sampling decision of false is set from trace header if no sample rate is available`() {
+        val logger = mock<ILogger>()
+        val propagationContext = PropagationContext.fromHeaders(
+            logger,
+            SentryTraceHeader(SentryId(), SpanId(), false).value,
+            "sentry-trace_id=a,sentry-transaction=sentryTransaction"
+        )
+        val context = TransactionContext.fromPropagationContext(propagationContext)
         assertNull(context.sampled)
         assertNull(context.profileSampled)
         assertFalse(context.parentSampled!!)
@@ -52,10 +60,14 @@ class TransactionContextTest {
     }
 
     @Test
-    fun `when context is created from trace header and baggage header, parent sampling decision of true is set from trace header`() {
-        val traceHeader = SentryTraceHeader(SentryId(), SpanId(), true)
-        val baggageHeader = Baggage.fromHeader("sentry-trace_id=a,sentry-transaction=sentryTransaction,sentry-sample_rate=0.3")
-        val context = TransactionContext.fromSentryTrace("name", TransactionNameSource.CUSTOM, "op", traceHeader, baggageHeader, null)
+    fun `when context is created from propagation context, parent sampling decision of true is set from trace header`() {
+        val logger = mock<ILogger>()
+        val propagationContext = PropagationContext.fromHeaders(
+            logger,
+            SentryTraceHeader(SentryId(), SpanId(), true).value,
+            "sentry-trace_id=a,sentry-transaction=sentryTransaction,sentry-sample_rate=0.3"
+        )
+        val context = TransactionContext.fromPropagationContext(propagationContext)
         assertNull(context.sampled)
         assertNull(context.profileSampled)
         assertTrue(context.parentSampled!!)
@@ -63,10 +75,14 @@ class TransactionContextTest {
     }
 
     @Test
-    fun `when context is created from trace header and baggage header, parent sampling decision of true is set from trace header if no sample rate is available`() {
-        val traceHeader = SentryTraceHeader(SentryId(), SpanId(), true)
-        val baggageHeader = Baggage.fromHeader("sentry-trace_id=a,sentry-transaction=sentryTransaction")
-        val context = TransactionContext.fromSentryTrace("name", TransactionNameSource.CUSTOM, "op", traceHeader, baggageHeader, null)
+    fun `when context is created from propagation context, parent sampling decision of true is set from trace header if no sample rate is available`() {
+        val logger = mock<ILogger>()
+        val propagationContext = PropagationContext.fromHeaders(
+            logger,
+            SentryTraceHeader(SentryId(), SpanId(), true).value,
+            "sentry-trace_id=a,sentry-transaction=sentryTransaction"
+        )
+        val context = TransactionContext.fromPropagationContext(propagationContext)
         assertNull(context.sampled)
         assertNull(context.profileSampled)
         assertTrue(context.parentSampled!!)